Add option chains (#37)
* use Decimal; fix typing in __post_init__; remove from_dict

* add options chains, basic and nested

* fix tests (apparently StrEnum is 3.11)

+++ tastytrade/metrics.py
@@ -1,4 +1,5 @@
 from datetime import date
+from decimal import Decimal
 from typing import Any, TypedDict
 
 import requests
@@ -8,20 +9,20 @@ from tastytrade.utils import validate_response
 
 DividendInfo = TypedDict('DividendInfo', {
     'occurred-date': date,
-    'amount': float
+    'amount': Decimal
 }, total=False)
 EarningsInfo = TypedDict('EarningsInfo', {
     'occurred-date': date,
-    'eps': float
+    'eps': Decimal
 }, total=False)
 MarketMetricInfo = TypedDict('MarketMetricInfo', {
     'symbol': str,
-    'implied-volatility-index': float,
-    'implied-volatility-index-5-day-change': float,
-    'implied-volatility-rank': float,
-    'implied-volatility-percentile': float,
-    'liquidity': float,
-    'liquidity-rank': float,
+    'implied-volatility-index': Decimal,
+    'implied-volatility-index-5-day-change': Decimal,
+    'implied-volatility-rank': Decimal,
+    'implied-volatility-percentile': Decimal,
+    'liquidity': Decimal,
+    'liquidity-rank': Decimal,
     'liquidity-rating': int,
     'option-expiration-implied-volatilities': list[dict[str, Any]]
 }, total=False)
